apps/api/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Cache
redis>=5.0.0

# HTTP client (Binance API)
httpx>=0.27.0

# Config & validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Auth & security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# Scheduler
apscheduler>=3.10.4

# Logging
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
